requirements.txt
streamlit
PyPDF2
reportlab
pillow
